Cambios en modulo de resoluciones

+++ app/Http/Controllers/ResolutionController.php
@@ -8,6 +8,8 @@ use App\Models\Resolution\FileResolution;
 use App\Http\Requests\Resolution\ResolutionStoreRequest;
 use App\Http\Requests\Resolution\ResolutionUpdateRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
+use Exception;
 use Storage;
 use Auth;
 use Str;
@@ -20,6 +22,10 @@ class ResolutionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create(Statement $statement) {
+        if (Auth::user()->hasRole('Cliente') && !is_null($statement['company']) && $statement['company']->user_id!=Auth::id()) {
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Consulta fallida', 'msg' => 'Acceso no permitido.']);
+        }
+
         return view('admin.resolutions.create', compact('statement'));
     }
 
@@ -30,6 +36,10 @@ class ResolutionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(ResolutionStoreRequest $request, Statement $statement) {
+        if (Auth::user()->hasRole('Cliente') && !is_null($statement['company']) && $statement['company']->user_id!=Auth::id()) {
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Registro fallido', 'msg' => 'Acceso no permitido.']);
+        }
+
         $data=array('name' => request('name'), 'description' => request('description'), 'date' => request('date'), 'statement_id' => $statement->id);
         $resolution=Resolution::create($data);
 
@@ -67,7 +77,11 @@ class ResolutionController extends Controller
      */
     public function show(Statement $statement, Resolution $resolution) {
         if (Auth::user()->hasRole('Cliente') && !is_null($statement['company']) && $statement['company']->user_id!=Auth::id()) {
-            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Consulta fallida', 'msg' => 'Acceso no permitido.']);
+        }
+
+        if (Auth::user()->hasRole('Cliente') && $statement->id!=$resolution->statement_id) {
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Consulta fallida', 'msg' => 'Acceso no permitido.']);
         }
 
         return view('admin.resolutions.show', compact('statement', 'resolution'));
@@ -84,6 +98,10 @@ class ResolutionController extends Controller
             return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
         }
 
+        if (Auth::user()->hasRole('Cliente') && $statement->id!=$resolution->statement_id) {
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
+        }
+
         return view('admin.resolutions.edit', compact("statement", 'resolution'));
     }
 
@@ -96,6 +114,10 @@ class ResolutionController extends Controller
      */
     public function update(ResolutionUpdateRequest $request, Statement $statement, Resolution $resolution) {
         if (Auth::user()->hasRole('Cliente') && !is_null($statement['company']) && $statement['company']->user_id!=Auth::id()) {
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
+        }
+
+        if (Auth::user()->hasRole('Cliente') && $statement->id!=$resolution->statement_id) {
             return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
         }
 
@@ -121,6 +143,10 @@ class ResolutionController extends Controller
             return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
         }
 
+        if (Auth::user()->hasRole('Cliente') && $statement->id!=$resolution->statement_id) {
+            return redirect()->route('statements.index')->with(['alert' => 'lobibox', 'type' => 'error', 'title' => 'Edición fallida', 'msg' => 'Acceso no permitido.']);
+        }
+
         $resolution->delete();
         if ($resolution) {
             return redirect()->route('statements.show', ['statement' => $statement->slug])->with(['alert' => 'sweet', 'type' => 'success', 'title' => 'Eliminación exitosa', 'msg' => 'El caso ha sido eliminado exitosamente.']);
@@ -142,13 +168,20 @@ class ResolutionController extends Controller
     }
 
     public function fileEdit(Request $request, Statement $statement, Resolution $resolution) {
+        if (Auth::user()->hasRole('Cliente') && !is_null($statement['company']) && $statement['company']->user_id!=Auth::id()) {
+            return response()->json(['status' => false], 403);
+        }
+
+        if (Auth::user()->hasRole('Cliente') && $statement->id!=$resolution->statement_id) {
+            return response()->json(['status' => false], 403);
+        }
+
         if ($request->hasFile('file')) {
             $file=$request->file('file');
             $name=time().'_'.Str::slug(substr($file->getClientOriginalName(), 0, "-".strlen($file->getClientOriginalExtension())), "-").".".$file->getClientOriginalExtension();
             $file->move(public_path().'/admins/files/statements/', $name);
 
-            $file=FileResolution::create(['name' => $name, 'resolution_id' => $resolution->id]);
-            if ($file) {
+            try {
                 $path='/';
                 $recursive=false;
                 $contents=collect(Storage::disk('google')->listContents($path, $recursive));
@@ -159,8 +192,14 @@ class ResolutionController extends Controller
                 $subdirectory=$contents->where('type', '=', 'dir')->where('filename', '=', $statement->slug)->first();
 
                 $filePath=public_path('admins/files/statements/'.$name);
-                Storage::disk('google')->put($directory['path'].'/'.$subdirectory['path'].'/'.$name, fopen($filePath, 'r+'));
+                Storage::disk('google')->put($directory['path'].'/'.$subdirectory['path'].'/'.$name, fopen($filePath, 'r+')); 
+            } catch (Exception $e) {
+                Log::error("Google API Exception: ".$e->getMessage());
+                return response()->json(['code' => 500, 'status' => 'error', 'message' => 'Ocurrió un error durante el proceso, intente nuevamente.'], 500);
+            }
 
+            $file=FileResolution::create(['name' => $name, 'resolution_id' => $resolution->id]);
+            if ($file) {
                 return response()->json(['status' => true, 'name' => $name, 'url' => asset('/admins/files/statements/'.$name), 'slug' => $statement->slug]);
             }
         }
@@ -169,6 +208,14 @@ class ResolutionController extends Controller
     }
 
     public function fileDestroy(Request $request, Statement $statement, Resolution $resolution) {
+        if (Auth::user()->hasRole('Cliente') && !is_null($statement['company']) && $statement['company']->user_id!=Auth::id()) {
+            return response()->json(['status' => false], 403);
+        }
+
+        if (Auth::user()->hasRole('Cliente') && $statement->id!=$resolution->statement_id) {
+            return response()->json(['status' => false], 403);
+        }
+
         $file=FileResolution::where('resolution_id', $resolution->id)->where('name', request('url'))->first();
         if (!is_null($file)) {
             $file->delete();
@@ -178,19 +225,23 @@ class ResolutionController extends Controller
                     unlink(public_path().'/admins/files/statements/'.request('url'));
                 }
 
-                $path='/';
-                $recursive=false;
-                $contents=collect(Storage::disk('google')->listContents($path, $recursive));
-                $directory=$contents->where('type', '=', 'dir')->where('filename', '=', $statement['company']->slug)->first();
+                try {
+                    $path='/';
+                    $recursive=false;
+                    $contents=collect(Storage::disk('google')->listContents($path, $recursive));
+                    $directory=$contents->where('type', '=', 'dir')->where('filename', '=', $statement['company']->slug)->first();
 
-                $path='/'.$directory['path'].'/';
-                $contents=collect(Storage::disk('google')->listContents($path, $recursive));
-                $subdirectory=$contents->where('type', '=', 'dir')->where('filename', '=', $statement->slug)->first();
+                    $path='/'.$directory['path'].'/';
+                    $contents=collect(Storage::disk('google')->listContents($path, $recursive));
+                    $subdirectory=$contents->where('type', '=', 'dir')->where('filename', '=', $statement->slug)->first();
 
-                $path='/'.$directory['path'].'/'.$subdirectory['path'].'/';
-                $contents=collect(Storage::disk('google')->listContents($path, $recursive));
-                $file_drive=$contents->where('type', '=', 'file')->where('filename', '=', pathinfo($file->name, PATHINFO_FILENAME))->where('extension', '=', pathinfo($file->name, PATHINFO_EXTENSION))->first();
-                Storage::disk('google')->delete($directory['path'].'/'.$subdirectory['path'].'/'.$file_drive['path']);
+                    $path='/'.$directory['path'].'/'.$subdirectory['path'].'/';
+                    $contents=collect(Storage::disk('google')->listContents($path, $recursive));
+                    $file_drive=$contents->where('type', '=', 'file')->where('filename', '=', pathinfo($file->name, PATHINFO_FILENAME))->where('extension', '=', pathinfo($file->name, PATHINFO_EXTENSION))->first();
+                    Storage::disk('google')->delete($directory['path'].'/'.$subdirectory['path'].'/'.$file_drive['path']);
+                } catch (Exception $e) {
+                    Log::error("Google API Exception: ".$e->getMessage());
+                }
 
                 return response()->json(['status' => true]);
             }
